pwr_zpi_backend/requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
coverage==6.1.2
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
dj-rest-auth==2.1.11
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.8.0
django-ipware==4.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-writable-nested==0.6.3
h11==0.12.0
idna==3.2
joblib==1.1.0
jwcrypto==1.0
libretranslatepy==2.1.1
lxml==4.6.4
mysqlclient==2.0.3
nltk==3.6.5
oauthlib==3.1.1
outcome==1.1.0
Pillow==8.4.0
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==21.0.0
python3-openid==3.2.0
pytz==2021.1
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.1
textblob==0.17.1
tqdm==4.62.3
translate==3.6.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.6
var-dump==1.2
wrapt==1.12.1
wsproto==1.0.0
